Ensure all colliders are known species or inerts
When reading in primary reaction libraries or seed mechanisms, third body
reactions can define collider efficiencies. If these are not declared species
nor are they declared inerts, they will be added as inerts.

This is not yet working - in fact it doesn't yet compile

For now, just ensure that any colliders in your seed mechanismor primary
reaction library, are defined either in the appropriate species.txt file or as
inerts in your condition file. (not sure what happens if they appear in both!)