Bug fix: Unreactive and/or ConstantConcentration flags for species
If a user defined a species as "Unreactive" or having
"ConstantConcentration," RMG would add multiple instances of the same species
to the speciesStatus list.  This is the list RMG uses to communicate with
the ODE solver, so indexing is critical.

For example,
CH4 (mol/cm3) 1 ConstantConcentration
1 C 0

RMG would add the "core" species "CH4" to the speciesStatus list with zero
flux and "1" concentration.  Upon reading the ConstantConcentration flag,
RMG would do the exact same thing, thus the additional species, and incorrect
indexing when passing information to the ODE solver.

Closes #123