Added Logger class for writing log messages.
The Logger class provides a few features not previously available in RMG: the
ability to vary the verbosity of the output log messages, and the ability to
write to multiple targets (e.g. the console and a file on disk). Ultimately
every System.out.print*() and System.err.print*() will be replaced with one of
the methods in this class.

Java has a few existing packages for logging, but the one in the standard
library (i.e. included with all recent Java releases) does not seem to be very
popular. Furthermore, I'm more familiar with the logging package in the Python
standard library; this class is based on that functionality.