Refinements to the RMG header.
Most notable is that the header is now printed via the logger.logHeader()
method. This allows multiple programs to print the header if so desired.

I moved a few things around in the header and deleted a few superfluous blank
lines. However, I have *not* changed the copyright information or the list of
authors.

I've also taken advantage of String.format() to clean up the right-hand side
"#" marks so that they form a nice box as originally intended.