Now redirecting stderr to stdout for log messages.
This is so that the error messages are always placed in the context of when
they occur. In particular, GUIs like Eclipse and NetBeans will semi-randomly
mix stderr and stdout in their output unless this is done. Note that the end
user will no longer be able to separate their error/warning messages from the
rest of the log.