Get merged code to compile (by removing some of it).
One of the changes on the master branch that we just merged in
was to remove ReactionTemplate attribute from Reaction objects
(only TemplateReactions have a ReactionTemplate now).
The check to see if it was from an H_Abstraction family
therefore broke. As it was doing nothing anyway, we have
just commented it out.